fix nip50 search input autofocus
